+++ CHANGELOG.md
@@ -7,7 +7,7 @@ dev
 0.10 (2023-09-10)
 ----------------
 
-- `notify-send` is now called with `--category im.received`
+* `notify-send` is now called with `--category im.received`
   ([#27](https://github.com/s3rvac/weechat-notify-send/issues/27)).
 * Dropped official support for Python 3.5, 3.6, and 3.7. as [they are
   EOL](https://endoflife.date/python).